extracted an axios instance from UserInfo. working on getting UserInfo to update from the backend. WIP.
